requirements.txt
python-pptx==0.6.19
openpyxl==2.6.4
pandas=1.1.5